move merge tests to add_book.match
